style: Moving more code around
